# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "FDP Kasapin",
        "description": "This API contains all allowable functionalities to be available internally to other resources & apps.You need to request a token to access this resource.",
        "contact": {
            "email": "jdavis@grameenfoundation.org"
        },
        "license": {
            "name": "Grameen Foundation",
            "url": "http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/user/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Login as field Coordinator",
                "operationId": "api.auth.user.login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "login email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "login password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation."
                    },
                    "401": {
                        "description": "Invalid Credentials."
                    },
                    "500": {
                        "description": "Could not create token"
                    }
                }
            }
        },
...
    },
...
}